test: move spec conditionals (ifit, ifdescribe)
