+++ spec/api-autoupdater-darwin-spec.ts
@@ -7,7 +7,7 @@ import * as os from 'os';
 import * as path from 'path';
 import * as psList from 'ps-list';
 import { AddressInfo } from 'net';
-import { ifdescribe, ifit } from './lib/spec-helpers';
+import { ifdescribe, ifit } from './lib/spec-conditional';
 import * as uuid from 'uuid';
 import { systemPreferences } from 'electron';
 import { fixturePath } from './lib/fixtures';
