+++ spec/api-crash-reporter-spec.ts
@@ -3,7 +3,8 @@ import * as childProcess from 'child_process';
 import * as http from 'http';
 import * as Busboy from 'busboy';
 import * as path from 'path';
-import { ifdescribe, ifit, defer, startRemoteControlApp, repeatedly, listen } from './lib/spec-helpers';
+import { defer, startRemoteControlApp, repeatedly, listen } from './lib/spec-helpers';
+import { ifdescribe, ifit } from './lib/spec-conditional';
 import { app } from 'electron/main';
 import { crashReporter } from 'electron/common';
 import { EventEmitter } from 'events';
