+++ spec/api-global-shortcut-spec.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import { globalShortcut } from 'electron/main';
-import { ifdescribe } from './lib/spec-helpers';
+import { ifdescribe } from './lib/spec-conditional';
 
 ifdescribe(process.platform !== 'win32')('globalShortcut module', () => {
   beforeEach(() => {
