+++ spec/api-notification-dbus-spec.ts
@@ -9,7 +9,7 @@
 import { expect } from 'chai';
 import * as dbus from 'dbus-native';
 import { app } from 'electron/main';
-import { ifdescribe } from './lib/spec-helpers';
+import { ifdescribe } from './lib/spec-conditional';
 import { promisify } from 'util';
 
 const skip = process.platform !== 'linux' ||
