+++ spec/api-power-monitor-spec.ts
@@ -8,7 +8,8 @@
 // python-dbusmock.
 import { expect } from 'chai';
 import * as dbus from 'dbus-native';
-import { ifdescribe, startRemoteControlApp } from './lib/spec-helpers';
+import { startRemoteControlApp } from './lib/spec-helpers';
+import { ifdescribe } from './lib/spec-conditional';
 import { promisify } from 'util';
 import { setTimeout } from 'timers/promises';
