+++ spec/api-subframe-spec.ts
@@ -2,7 +2,8 @@ import { expect } from 'chai';
 import * as http from 'http';
 import { closeWindow } from './lib/window-helpers';
 import { app, BrowserWindow, ipcMain } from 'electron/main';
-import { ifdescribe, listen } from './lib/spec-helpers';
+import { listen } from './lib/spec-helpers';
+import { ifdescribe } from './lib/spec-conditional';
 import { once } from 'events';
 import { emittedN } from './lib/events';
 import { fixturePath } from './lib/fixtures';
