+++ spec/api-tray-spec.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import { Menu, Tray } from 'electron/main';
 import { nativeImage } from 'electron/common';
-import { ifdescribe, ifit } from './lib/spec-helpers';
+import { ifdescribe, ifit } from './lib/spec-conditional';
 import * as path from 'path';
 
 describe('tray module', () => {
