+++ spec/api-web-frame-main-spec.ts
@@ -4,7 +4,8 @@ import * as path from 'path';
 import * as url from 'url';
 import { BrowserWindow, WebFrameMain, webFrameMain, ipcMain, app, WebContents } from 'electron/main';
 import { closeAllWindows } from './lib/window-helpers';
-import { defer, ifit, listen, waitUntil } from './lib/spec-helpers';
+import { defer, listen, waitUntil } from './lib/spec-helpers';
+import { ifit } from './lib/spec-conditional';
 import { once } from 'events';
 import { setTimeout } from 'timers/promises';
 import { emittedN } from './lib/events';
