+++ spec/asar-spec.ts
@@ -4,7 +4,8 @@ import * as url from 'url';
 import { Worker } from 'worker_threads';
 import { BrowserWindow, ipcMain } from 'electron/main';
 import { closeAllWindows } from './lib/window-helpers';
-import { getRemoteContext, ifdescribe, ifit, itremote, useRemoteContext } from './lib/spec-helpers';
+import { getRemoteContext, itremote, useRemoteContext } from './lib/spec-helpers';
+import { ifdescribe, ifit } from './lib/spec-conditional';
 import * as importedFs from 'fs';
 import { once } from 'events';
 import { fixtureFileURL, fixturePath } from './lib/fixtures';
