+++ spec/chromium-spec.ts
@@ -9,7 +9,8 @@ import * as url from 'url';
 import * as ChildProcess from 'child_process';
 import { EventEmitter, once } from 'events';
 import { promisify } from 'util';
-import { ifit, ifdescribe, defer, itremote, listen } from './lib/spec-helpers';
+import { defer, itremote, listen } from './lib/spec-helpers';
+import { ifdescribe, ifit } from './lib/spec-conditional';
 import { PipeTransport } from './pipe-transport';
 import * as ws from 'ws';
 import { setTimeout } from 'timers/promises';
