+++ spec/extensions-spec.ts
@@ -5,7 +5,8 @@ import * as http from 'http';
 import * as path from 'path';
 import * as fs from 'fs';
 import * as WebSocket from 'ws';
-import { ifit, listen } from './lib/spec-helpers';
+import { listen } from './lib/spec-helpers';
+import { ifit } from './lib/spec-conditional';
 import { once } from 'events';
 import { emittedN, emittedUntil } from './lib/events';
 import { fixturePath } from './lib/fixtures';
