+++ spec/lib/spec-conditional.ts
@@ -1,0 +1,7 @@
+export function ifit (cond: boolean): Mocha.TestFunction | Mocha.PendingTestFunction {
+  return cond ? it : it.skip;
+}
+
+export function ifdescribe (cond: boolean): Mocha.SuiteFunction | Mocha.PendingSuiteFunction {
+  return cond ? describe : describe.skip;
+}
