+++ spec/node-spec.ts
@@ -2,7 +2,8 @@ import { expect } from 'chai';
 import * as childProcess from 'child_process';
 import * as fs from 'fs';
 import * as util from 'util';
-import { getRemoteContext, ifdescribe, ifit, itremote, useRemoteContext } from './lib/spec-helpers';
+import { getRemoteContext, itremote, useRemoteContext } from './lib/spec-helpers';
+import { ifdescribe, ifit } from './lib/spec-conditional';
 import { webContents } from 'electron/main';
 import { EventEmitter } from 'stream';
 import { once } from 'events';
