+++ spec/spellchecker-spec.ts
@@ -5,7 +5,8 @@ import * as path from 'path';
 import * as fs from 'fs';
 import * as http from 'http';
 import { closeWindow } from './lib/window-helpers';
-import { ifit, ifdescribe, listen } from './lib/spec-helpers';
+import { listen } from './lib/spec-helpers';
+import { ifdescribe, ifit } from './lib/spec-conditional';
 import { once } from 'events';
 import { setTimeout } from 'timers/promises';
 import { fixturePath } from './lib/fixtures';
